+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Change Log
 4.2.3 (unreleased)
 --------------------------
 
+### Improvements
+
+- [trivial] Change the color names in the Stylesheet Keyword color palette in the color code panel to lower case.
+
+
 ### Fixes
 
 - Fix an issue that letters in the editor were drawn in wrong glyphs when updating the font under very specific conditions.